feat: support negative entries in CellIndex (unsigned int -> int)

+++ coupling/filtering/filters/FilterFromFunction.h
@@ -38,11 +38,11 @@ class coupling::FilterFromFunction : public coupling::FilterInterface<dim> {
 				throw std::runtime_error("ERROR: FilterFromFunction received nullptr as function pointer!");
 
 			//cast MaMiCo indexing to std::array
-			tarch::la::Vector<dim, unsigned int> mamicoIndex;
+			tarch::la::Vector<dim, int> mamicoIndex;
 			std::array<unsigned int, dim> stlIndex;
 			for(unsigned int i = 0; i < inputCellVector.size(); i++) {
 				//interpret position of cell in inputCellVector as linear local md-to-macro index, then convert it to vector
-				mamicoIndex = coupling::indexing::convertToVector<dim, IndexTrait::local, IndexTrait::md2macro>( {i} );
+				mamicoIndex = coupling::indexing::convertToVector<dim, IndexTrait::local, IndexTrait::md2macro>( { (int) i } );
 				for(unsigned int d = 0; d < dim; d++) stlIndex[d] = mamicoIndex[d];
 				_stlIndices.push_back(stlIndex);
 			}
